implementacion

+++ src/main/java/com/jsnunez/peliculas/domain/repository/ClienteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Cliente;
 
 public interface ClienteRepository  extends
-CrudRepository<Cliente,Short> {
+CrudRepository<Cliente,Long> {
 }
