+++ src/main/java/com/jsnunez/peliculas/domain/repository/PagoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Pago;
 
 public interface PagoRepository  extends
-CrudRepository<Pago,Short> {
+CrudRepository<Pago,Integer> {
 }
