+++ src/main/java/com/jsnunez/peliculas/domain/repository/PeliculaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Pelicula;
 
 public interface PeliculaRepository  extends
-CrudRepository<Pelicula,Short> {
+CrudRepository<Pelicula,Long> {
 }
